Implement class-factory + shim for hints extracting and remove the utils module
